[model] split npu patcher
